+++ config/async.php
@@ -2,6 +2,11 @@
 
 return [
     /*
+     * The PHP binary will be use in async processes.
+     */
+    'withBinary' => PHP_BINARY,
+
+    /*
      * Maximum concurrency async processes.
      */
     'concurrency' => 20,
@@ -15,6 +20,11 @@ return [
      * Sleep (micro-second) time when waiting async processes.
      */
     'sleepTime' => 50000,
+
+    /*
+     * Default output length of async processes.
+     */
+    'defaultOutputLength' => 1024 * 10,
 
     /*
      * An autoload script to boot composer autoload and Laravel application.